LPS-61095 move to modules
