+++ modules/portal/portal-search-test/src/testIntegration/java/com/liferay/portal/search/internal/test/TestSingleIndexerPostProcessor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.deploy.hot.bundle.indexerpostprocessorregistry;
+package com.liferay.portal.search.internal.test;
 
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.Document;
@@ -30,11 +30,12 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	immediate = true,
-	property = {"indexer.class.name=com.liferay.portal.model.Contact"},
+	property = {
+		"indexer.class.name=com.liferay.portlet.usersadmin.util.OrganizationIndexer"
+	},
 	service = IndexerPostProcessor.class
 )
-public class TestSingleEntityIndexerPostProcessor
-	implements IndexerPostProcessor {
+public class TestSingleIndexerPostProcessor implements IndexerPostProcessor {
 
 	@Override
 	public void postProcessContextBooleanFilter(
